🛠️  Adjusting redirect routes when the call is external of manager
